Extract shared CLI helpers (CliConnectionResolver + PlanAnalysisRunner)
AnalyzeCommand and QueryStoreCommand duplicated their connection/credential
resolution, server-metadata fetch, the parse->analyze->score pipeline, and
the json/text/both result-file writing. Several copies had drifted.

- CliConnectionResolver: IsAzureSqlDb, BuildServerConnection (auth-type
  resolution + SQL-credential existence check), and the non-fatal
  FetchServerMetadataAsync.
- PlanAnalysisRunner: Analyze (parse + analyze + score, branching on
  serverMetadata) and WriteResultFilesAsync (warningsOnly + json/text/both).

Both commands now share these. querystore picks up AnalyzeCommand's
fail-fast "no credential found" check (wrapped in the same graceful
try/catch) instead of proceeding to a cryptic downstream SQL error.
Offline analyze, live capture, and querystore behavior otherwise unchanged;
verified the offline analyze path end to end on a sample plan. 85 tests green.
